requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
bcrypt==4.0.1
certifi==2025.8.3
click==8.2.1
coverage==7.10.2
ecdsa==0.19.1
fastapi==0.116.1
graphql-core==3.2.6
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
lia-web==0.2.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2==2.9.10
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
strawberry-graphql==0.278.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
